fix: new login flow
